requirements.txt
torch==2.6.0
vllm==0.8.5
faiss-gpu==1.7.2
deepspeed==0.17.4
python-liquid==1.10.2
openai==1.86.0
tqdm==4.67.1
transformers==4.53.2
sentence-transformers==5.0.0
pyserini==1.0.0
tiktoken==0.9.0
datasets==3.6.0
deepspeed==0.17.4
